+++ ui/src/pages/admin/provider-form.tsx
@@ -1,4 +1,4 @@
-import { Save } from "@mui/icons-material";
+import { ArrowBackTwoTone, Save } from "@mui/icons-material";
 import {
   Container,
   TextField,
@@ -191,6 +191,9 @@ export const ProviderForm: React.FC = (): React.ReactElement => {
             alignItems: "center",
           }}
         >
+          <Button variant="text" sx={{ mr: 1 }} href="/admin">
+            <ArrowBackTwoTone />
+          </Button>
           <Typography variant="h4" sx={{ mr: 2 }}>
             {id ? `Edit Provider: ` : "New Provider"}
           </Typography>