+++ ui/src/pages/admin/user-form.tsx
@@ -2,6 +2,8 @@ import { FetchResult } from "@apollo/client";
 import {
   AdminPanelSettings,
   AdminPanelSettingsTwoTone,
+  ArrowBack,
+  ArrowBackTwoTone,
   LockResetTwoTone,
   Save,
   SupervisorAccountTwoTone,
@@ -224,6 +226,9 @@ export const UserForm: React.FC = (): React.ReactElement => {
             alignItems: "center",
           }}
         >
+          <Button variant="text" sx={{ mr: 1 }} href="/admin">
+            <ArrowBackTwoTone />
+          </Button>
           <Typography variant="h4" sx={{ mr: 2 }}>
             {id ? `Edit User: ` : "New User"}
           </Typography>